requirements.txt
PyMuPDF
streamlit
google-generativeai
pdf2image
fastapi 
uvicorn 
spacy
PyPDF2




